EOS-17417: [DTM0] Propagate E state for DTX.
The patch extends m0c and dix code to propagate
the Executed ("E" in the title) state from CAS request
to DTX, and it separtes the "completed" machinery (asts, callbacks)
into separate "executed" and "stable" on the m0c side.

Note1: the patch has a few TODO/XXX comments for the reviewers.
Some of them may need to be resolved before considering
this patch for merging.
Note2: the patch was tested with "--enable-dtm0" only.
It needs to be verified with "--disable-dtm0".